move block codec into separate file
